update on all QA data that is working

+++ apollo-api/src/main/kotlin/org/kebs/app/kotlin/apollo/api/handlers/QualityAssuranceInternalUserHandler.kt
@@ -748,7 +748,7 @@ class QualityAssuranceInternalUserHandler(
             validator.validate(body, errors)
             when {
                 errors.allErrors.isEmpty() -> {
-                    qaDaoServices.updatePermitSaveLabPDFSelectedDetails(permitID.toLong(), body)
+                    qaDaoServices.updatePermitSaveLabPDFSelectedDetails(permitID, body)
                         .let {
                             ok().body(it)
                         }
